Add a forwarding header
